+++ mne/io/bti/bti.py
@@ -1000,7 +1000,7 @@ class RawBTi(BaseRaw):
 def _make_bti_digitization(
         info, head_shape_fname, convert, use_hpi, bti_dev_t, dev_ctf_t):
 
-    from ...digitization._mess import _make_bti_dig_points
+    from ...digitization._utils import _make_bti_dig_points
 
     if head_shape_fname:
         logger.info('... Reading digitization points from %s' %
